# api/docs/swagger.yaml
definitions:
  models.Client:
    properties:
      fullname:
        type: string
      id:
        type: string
      phone:
        type: string
    type: object
  models.CreateClientBody:
    properties:
      fullname:
        type: string
      phone:
        type: string
    type: object
  models.CreateDriverBody:
    properties:
      fullname:
        type: string
      phone:
        type: string
    type: object
  models.CreateOrder:
    properties:
      clientId:
        type: string
      driverId:
        type: string
      status:
        type: string
    type: object
  models.Driver:
    properties:
      fullname:
        type: string
      id:
        type: string
      phone:
        type: string
    type: object
  models.Error:
    properties:
      message:
        type: string
    type: object
  models.FullOrder:
    properties:
      client:
        $ref: '#/definitions/models.Client'
      driver:
        $ref: '#/definitions/models.Driver'
      id:
        type: string
      status:
        type: string
    type: object
  models.ListOrders:
    properties:
      count:
        type: integer
      orders:
        items:
          $ref: '#/definitions/models.Order'
        type: array
    type: object
  models.Order:
    properties:
      clientId:
        type: integer
      driverId:
        type: integer
      id:
        type: string
      status:
        type: string
    type: object
  models.StandardErrorModel:
    properties:
      error:
        $ref: '#/definitions/models.Error'
    type: object
  models.UpdateOrder:
    properties:
      status:
        type: string
    type: object
info:
  contact: {}
paths:
  /v1/clients/:
    post:
      consumes:
      - application/json
      description: This API for creating a new client
      parameters:
      - description: clientCreateRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateClientBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: CreateClient
      tags:
      - client
  /v1/clients/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting client
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: DeleteClient
      tags:
      - client
    get:
      consumes:
      - application/json
      description: This API for getting client detail
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: GetClient
      tags:
      - client
    put:
      consumes:
      - application/json
      description: This API for updating client
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: clientUpdateRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateClientBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: UpdateClient
      tags:
      - client
  /v1/drivers/:
    post:
      consumes:
      - application/json
      description: This API for creating a new driver
      parameters:
      - description: driverCreateRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateDriverBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Driver'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: CreateDriver
      tags:
      - driver
  /v1/drivers/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting driver
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: DeleteDriver
      tags:
      - driver
    get:
      consumes:
      - application/json
      description: This API for getting driver detail
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Driver'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: GetDriver
      tags:
      - driver
    put:
      consumes:
      - application/json
      description: This API for updating driver
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: driverUpdateRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateDriverBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Driver'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: UpdateDriver
      tags:
      - driver
  /v1/order/{id}:
    get:
      consumes:
      - application/json
      description: This API for getting order detail
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FullOrder'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: GetOrder
      tags:
      - order
  /v1/orders/:
    post:
      consumes:
      - application/json
      description: This API for creating a new order
      parameters:
      - description: clientCreateRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateOrder'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FullOrder'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: CreateOrder
      tags:
      - order
  /v1/orders/{clientId}:
    get:
      consumes:
      - application/json
      description: This API for getting list of orders
      parameters:
      - description: ClientID
        in: path
        name: clientId
        required: true
        type: string
      - description: Page
        in: query
        name: page
        type: string
      - description: Limit
        in: query
        name: limit
        type: string
      - description: from
        in: query
        name: from
        type: string
      - description: to
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListOrders'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: ListOrders
      tags:
      - order
  /v1/orders/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting order
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: DeleteOrder
      tags:
      - order
    put:
      consumes:
      - application/json
      description: This API for updating order
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: orderUpdateRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateOrder'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FullOrder'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: UpdateOrder
      tags:
      - order
swagger: "2.0"